+++ components/Components/URLSearchParamsTest.php
@@ -742,7 +742,7 @@ JSON;
             ],
         ];
 
-        $params = URLSearchParams::fromParameters($parameters);
+        $params = URLSearchParams::fromPhpVariable($parameters);
         self::assertCount(4, $params);
         self::assertSame('bar baz', $params->get('filter[foo][0]'));
         self::assertSame('bar', $params->get('filter[bar][foo]'));
@@ -813,7 +813,7 @@ JSON;
 
         self::assertSame(
             URLSearchParams::fromAssociative($parameters)->toString(),
-            URLSearchParams::fromParameters($parameters)->toString(),
+            URLSearchParams::fromPhpVariable($parameters)->toString(),
         );
     }
 
@@ -824,7 +824,7 @@ JSON;
      */
     public function testFromParametersWithDifferentInput(object|array $data, string $expected): void
     {
-        self::assertSame($expected, URLSearchParams::fromParameters($data)->toString());
+        self::assertSame($expected, URLSearchParams::fromPhpVariable($data)->toString());
     }
 
     public static function providesParametersInput(): iterable